Built LevelByteStringSize.
It will ensure that all ByteStrings passing through it are the same
size. This is so I can use Zip on two flows of ByteStrings and keep
them in sync for use in CacheFile. I want to not block reading from
the network stream while computing the md5 of the file, but when the
md5 calc is complete switch to reading from the file stream.

Working on CacheFile design and building up components for its use.